Allow custom switch hotkeys
