+++ debian/changelog
@@ -1,3 +1,10 @@
+nkms (0.1.1) unstable; urgency=medium
+
+  * Allow custom switch hotkeys
+
+ -- Nicholas Yoder <nicholasyoder@proton.me>  Mon, 0 Jan 2025 00:00:00 +0000
+
+
 nkms (0.1) unstable; urgency=medium
 
   * Run daemon as a systemd service
