+++ etc/nkms/nkms.conf
@@ -8,3 +8,5 @@ server=
 [server]
 bind_address=0.0.0.0
 port=4777
+key1=125
+key2=41
