+++ nkms/core/settings.py
@@ -1,19 +1,6 @@
 from typing import Any
 
-from PyQt6.QtCore import QSettings, pyqtSignal
-from PyQt6.QtWidgets import (
-    QWidget,
-    QPushButton,
-    QVBoxLayout,
-    QHBoxLayout,
-    QStackedLayout,
-    QComboBox,
-    QLabel,
-    QGroupBox,
-    QGridLayout,
-    QLineEdit,
-    QSpinBox,
-)
+from PyQt6.QtCore import QSettings
 
 
 class NkmsSettings:
@@ -25,6 +12,8 @@ class NkmsSettings:
         self.client_port = "4777"
         self.server_address = "0.0.0.0"
         self.server_port = "4777"
+        self.server_key1 = 125
+        self.server_key2 = 41
         # Load saved values
         self.load()
 
@@ -35,6 +24,8 @@ class NkmsSettings:
         self.client_port = int(self.settings.value("client/port", self.client_port))
         self.server_address = self.settings.value("server/bind_address", self.server_address)
         self.server_port = int(self.settings.value("server/port", self.server_port))
+        self.server_key1 = int(self.settings.value("server/key1", self.server_key1))
+        self.server_key2 = int(self.settings.value("server/key2", self.server_key2))
 
     def save(self):
         for key, value in self.get_settings_dict():
@@ -48,98 +39,10 @@ class NkmsSettings:
             "client/port": self.client_port,
             "server/bind_address": self.server_address,
             "server/port": self.server_port,
+            "server/key1": self.server_key1,
+            "server/key2": self.server_key2,
         }
 
     def save_settings_dict(self, settings_dict: dict[str, Any]) -> None:
         for key, value in settings_dict.items():
             self.settings.setValue(key, value)
-
-
-class SettingsWindow(QWidget):
-
-    settings_saved = pyqtSignal(dict)
-
-    def __init__(self):
-        super().__init__()
-        self.settings = NkmsSettings()
-
-        self.setWindowTitle("NKMS Settings")
-
-        base_layout = QVBoxLayout()
-
-        mode_layout = QHBoxLayout()
-        mode_label = QLabel("Mode:")
-        mode_layout.addWidget(mode_label)
-        self.mode_select = QComboBox()
-        self.mode_select.addItem("Client")
-        self.mode_select.addItem("Server")
-        mode_layout.addWidget(self.mode_select)
-        mode_layout.addStretch(1)
-        base_layout.addLayout(mode_layout)
-
-        base_layout.addSpacing(5)
-
-        self.stacked_layout = QStackedLayout()
-
-        client_gbox = QGroupBox("Client Settings")
-        client_base_layout = QGridLayout()
-        client_base_layout.addWidget(QLabel("Server:"), 0, 0)
-        self.client_server_input = QLineEdit()
-        client_base_layout.addWidget(self.client_server_input, 0, 1)
-        client_base_layout.addWidget(QLabel("Port:"), 1, 0)
-        self.client_port_input = QSpinBox()
-        self.client_port_input.setRange(0, 65535)
-        client_base_layout.addWidget(self.client_port_input, 1, 1)
-        client_gbox.setLayout(client_base_layout)
-        client_base_layout.addWidget(QLabel("Note: Next higher port will also be used."), 2, 0, 2, 2)
-        self.stacked_layout.addWidget(client_gbox)
-
-        server_gbox = QGroupBox("Server Settings")
-        server_base_layout = QGridLayout()
-        server_base_layout.addWidget(QLabel("Bind Address:"), 0, 0)
-        self.server_address_input = QLineEdit()
-        server_base_layout.addWidget(self.server_address_input, 0, 1)
-        server_base_layout.addWidget(QLabel("Port:"), 1, 0)
-        self.server_port_input = QSpinBox()
-        self.server_port_input.setRange(0, 65535)
-        server_base_layout.addWidget(self.server_port_input, 1, 1)
-        server_gbox.setLayout(server_base_layout)
-        server_base_layout.addWidget(QLabel("Note: Next higher port will also be used."), 2, 0, 2, 2)
-        self.stacked_layout.addWidget(server_gbox)
-
-        base_layout.addLayout(self.stacked_layout)
-
-        base_layout.addSpacing(5)
-
-        buttons_layout = QHBoxLayout()
-        buttons_layout.addStretch(1)
-        close_bt = QPushButton("Close")
-        buttons_layout.addWidget(close_bt)
-        apply_bt = QPushButton("Save")
-        buttons_layout.addWidget(apply_bt)
-        buttons_layout.addStretch(1)
-        base_layout.addLayout(buttons_layout)
-
-        self.setLayout(base_layout)
-
-        self.mode_select.activated.connect(self.stacked_layout.setCurrentIndex)
-        close_bt.clicked.connect(self.close)
-        apply_bt.clicked.connect(self.apply_settings)
-
-        self.load_settings()
-
-    def load_settings(self):
-        self.mode_select.setCurrentText(self.settings.mode)
-        self.client_server_input.setText(self.settings.client_server)
-        self.client_port_input.setValue(self.settings.client_port)
-        self.server_address_input.setText(self.settings.server_address)
-        self.server_port_input.setValue(self.settings.server_port)
-        self.stacked_layout.setCurrentIndex(self.mode_select.currentIndex())
-
-    def apply_settings(self):
-        self.settings.mode = self.mode_select.currentText()
-        self.settings.client_server = self.client_server_input.text()
-        self.settings.client_port = self.client_port_input.text()
-        self.settings.server_address = self.server_address_input.text()
-        self.settings.server_port = self.server_port_input.text()
-        self.settings_saved.emit(self.settings.get_settings_dict())
